Split benchmarks into multiple files
