+++ crates/forge/benches/my_bench/mod.rs
@@ -1,1 +1,4 @@
-pub mod my_benchmark;
+pub mod benchmark;
+mod collect_tests;
+mod compile_tests;
+mod declare_deploy_and_interact;
